+++ json_test.go
@@ -17,7 +17,7 @@ func TestJSONParse(t *testing.T) {
 	if _, err := v8go.JSONParse(nil, "{}"); err == nil {
 		t.Error("expected error but got <nil>")
 	}
-	ctx, _ := v8go.NewContext()
+	ctx := v8go.NewContext()
 	defer ctx.Isolate().Dispose()
 	defer ctx.Close()
 	_, err := v8go.JSONParse(ctx, "{")
@@ -34,7 +34,7 @@ func TestJSONParse(t *testing.T) {
 func TestJSONStringify(t *testing.T) {
 	t.Parallel()
 
-	ctx, _ := v8go.NewContext()
+	ctx := v8go.NewContext()
 	defer ctx.Isolate().Dispose()
 	defer ctx.Close()
 	if _, err := v8go.JSONStringify(ctx, nil); err == nil {
@@ -43,7 +43,7 @@ func TestJSONStringify(t *testing.T) {
 }
 
 func ExampleJSONParse() {
-	ctx, _ := v8go.NewContext()
+	ctx := v8go.NewContext()
 	defer ctx.Isolate().Dispose()
 	defer ctx.Close()
 	val, _ := v8go.JSONParse(ctx, `{"foo": "bar"}`)
@@ -53,7 +53,7 @@ func ExampleJSONParse() {
 }
 
 func ExampleJSONStringify() {
-	ctx, _ := v8go.NewContext()
+	ctx := v8go.NewContext()
 	defer ctx.Isolate().Dispose()
 	defer ctx.Close()
 	val, _ := v8go.JSONParse(ctx, `{