+++ promise_test.go
@@ -15,7 +15,7 @@ func TestPromiseFulfilled(t *testing.T) {
 
 	iso := v8go.NewIsolate()
 	defer iso.Dispose()
-	ctx, _ := v8go.NewContext(iso)
+	ctx := v8go.NewContext(iso)
 	defer ctx.Close()
 
 	if _, err := v8go.NewPromiseResolver(nil); err == nil {
@@ -65,7 +65,7 @@ func TestPromiseRejected(t *testing.T) {
 
 	iso := v8go.NewIsolate()
 	defer iso.Dispose()
-	ctx, _ := v8go.NewContext(iso)
+	ctx := v8go.NewContext(iso)
 	defer ctx.Close()
 
 	res2, _ := v8go.NewPromiseResolver(ctx)
@@ -115,7 +115,7 @@ func TestPromiseThenPanic(t *testing.T) {
 
 	iso := v8go.NewIsolate()
 	defer iso.Dispose()
-	ctx, _ := v8go.NewContext(iso)
+	ctx := v8go.NewContext(iso)
 	defer ctx.Close()
 
 	res, _ := v8go.NewPromiseResolver(ctx)